Fix simpler conditional cases
This feels kind of gross and a "where will it end" kind of situation.

For example we could filter out `begin`, `rescue`, etc and it still wouldn't
be enough for something like this:

```
      begin
        v = find_version_from_server
      rescue
        puts "something bad happened"
      ensure
        v = "1.0"
      end

      gem 'super-secret-internal', v

      def find_version_from_server
        class Finder < Other
          def initialize
            super
          end
        end
        Finder.go!
      end
```

I'm kind of ok with return UNKNOWN for `Finder.go!`, but I could imagine
a version of this that defines a `gem` method and eval's it.  But that
would put a crimp on any future plans of hosting this tool since we'd be
doing RCE all over the place.